Update 06_StructuredConcurrency.md
It is a thought sentence. "We always start new coroutines inside a scope." sounds like all we do in scope is starting new coroutines. It is also not true that all coroutines start in some scope. `runBlocking` does not need that, as well as primitives functions and builders from external libraries.