Add explanation for external modifier
I think it would be helpful to have an explanation / link to the docs for the external modifier - happy for it to be reworded if you can think of something better. 

I haven't really used KotlinJS before, so I wasn't familiar with it. As I was going through the tutorial, I left off the modifier in one place because I wasn't sure if it was really necessary. That led to a not-very-helpful error message and a lengthy debugging session until i figured out that the error was caused by the missing modifier...